requirements/base.txt
# Core Django
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Task Queue
celery==5.3.6
redis==5.0.3
kombu==5.3.4
django-celery-beat==2.6.0

# Authentication & Security
PyJWT==2.8.0
cryptography==42.0.5

# Utilities
Pillow==10.3.0
phonenumbers==8.13.35
django-phonenumber-field==7.3.0
python-decouple==3.8
pytz==2024.1

# HTTP
requests==2.31.0
httpx==0.27.0

# Logging
structlog==24.1.0

# API documentation
drf-spectacular==0.27.1

# Geo / Spatial
geopy==2.4.1
